requirements.txt
sentence-transformers
faiss-cpu
transformers
torch
requests
llama-index
llama-index-readers-file
